doc-automation/scripts/requirements.txt
python-pptx>=0.6.21
pandas>=2.0.0
matplotlib>=3.7.0
openpyxl>=3.1.0
python-docx>=1.0.0
pdfplumber>=0.10.0
beautifulsoup4>=4.12.0
